requirements.txt
Flask
Pillow
python-dotenv
flask-cors
gunicorn
numpy